Add initial server config notes and setup in CL
I added a file documenting the process I used to set up the PostgreSQL
server, for posterity (or me) if we have to migrate. I added a new
dependancy, cl+ssl for connecting to the PostgreSQL server with
encryption enabled, and a quick line in olduvai.lisp to show example
code for making the connection. I still need to iron out the SSL
certification process, probably by generating a self-signed cert for
the server and including it in the repository.