Add the first version of the helper program, and update the README.
I switched to the Go programming language for this project, and
documented my reasons for this in the README. I also added the first
version of the helper program, stored in "olduvai.go" (eventually we
will rename it and possibly separate it into different
components). Right now the program includes a simple "database" data
type that I made for storing and manipulating our data before sending
it to the server. Importing a CSV file from Microsoft Access works
fine already.