Improve type inference and add hasDifferentValues field.
"hasDifferentValues" is a useful indicator of whether a variable is
going to be "useful" for generating visuals. Needs to be updated to
catch whitespace strings of different lengths.

Type inference is improved, but I need to fix it to default to string
for cases like this: "N" "N" "0" instead of integer.