Add showReference function.
For now, if the user requires more info about a term, we can
automatically link to its location on http://rs.tdwg.org/dwc/terms/
This includes the definition, examples, and comments about the
term. In the future we might try to programmatically import this info,
but currently that would involve web scraping (which is fragile), so
I'm going to wait and ask the DWC team about it when I submit
DWCHelper to them for review.