Add getAliases() function and start aliases.csv file.
getAliases is a useful way to detect probable DWC terms in a database
based off of the term names and the aliases file we provide. Right now
the alias file only contains three aliases - I'll need external help
to make it more complete.

I need to add the "detect unused terms" feature, then all the pieces
are ready to build and finish the actual user-interface part of the
program.